Git Ignore and starter content
Modified .gitignore to ignore everything in the StarterContent.  The idea is that everything offered there should be read only, and modifications should be copied and moved out of the startercontent subfolder.
Must go to Add new on the content browser and add starter content back in, then the git ignore will work as expected.